update: refactor buttonPanel
